Task-2. Corrected amplification realization. Still need to move amplifier to linearity function
(now the Fourier transform is performed once again).
Added modifications to Storage module.
Added the ability to write the simulation result to a file, with fiber parameters written as a comment.
Added ability to draw animation, based on storage data.
Added notebooks as an examples of reading and drawing animation.